Fix trivy failure
